+++ conan/tools/cmake/cmake.py
@@ -1,8 +1,8 @@
 import os
 
 from conan.tools.build import build_jobs
-from conan.tools.cmake.cmake_initial_cache import update_cmake_initial_cache, get_initial_cache_path
-from conan.tools.cmake.presets import load_cmake_presets, get_configure_preset
+from conan.tools.cmake.presets import load_cmake_presets, get_configure_preset, \
+    get_initial_cache_path
 from conan.tools.cmake.utils import is_multi_configuration
 from conan.tools.files import chdir, mkdir
 from conan.tools.microsoft.msbuild import msbuild_verbosity_cmd_line_arg
@@ -61,27 +61,6 @@ class CMake(object):
         self._presets_cache_variables = configure_preset["cacheVariables"]
         self._cmake_program = "cmake"  # Path to CMake should be handled by environment
 
-    @property
-    def _cache_variables(self):
-        """
-        Get all the variables that should be passed the CLI as ``cmake -Dxxxxx -Dyyyyy ....``
-        """
-        cache_variables = {}
-        generator_folder = self._conanfile.generators_folder
-        if self._toolchain_file:
-            if os.path.isabs(self._toolchain_file):
-                toolchain_file = self._toolchain_file
-            else:
-                toolchain_file = os.path.join(generator_folder, self._toolchain_file)
-            cache_variables["CMAKE_TOOLCHAIN_FILE"] = toolchain_file.replace("\\", "/")
-
-        if self._conanfile.package_folder:
-            pkg_folder = self._conanfile.package_folder.replace("\\", "/")
-            cache_variables["CMAKE_INSTALL_PREFIX"] = pkg_folder
-
-        cache_variables.update(self._presets_cache_variables)
-        return cache_variables
-
     def configure(self, variables=None, build_script_folder=None):
         cmakelist_folder = self._conanfile.source_folder
         if build_script_folder:
@@ -90,12 +69,17 @@ class CMake(object):
         build_folder = self._conanfile.build_folder
         mkdir(self._conanfile, build_folder)
 
-        arg_list = [self._cmake_program]
-        # Running the initial cache file
-        cache_variables = self._cache_variables
-        cache_variables.update(variables or {})
-        update_cmake_initial_cache(self._conanfile, cache_variables)
-        arg_list.append(f'-C "{get_initial_cache_path(self._conanfile)}"')
+        arg_list = [
+            self._cmake_program,
+            f'-C "{get_initial_cache_path(self._conanfile)}"',
+        ]
+
+        if self._conanfile.package_folder:
+            pkg_folder = self._conanfile.package_folder.replace("\\", "/")
+            arg_list.append(f'-DCMAKE_INSTALL_PREFIX="{pkg_folder}"')
+
+        if variables:
+            arg_list.extend([f'-D{k}="{v}"' for k, v in variables.items()])
 
         if self._generator:
             arg_list.append('-G "{}"'.format(self._generator))
